refactor: restructure user context and hooks for improved organization
