+++ src/hooks/useUser.js
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import { UserContext } from "../Context/UserContext/UserContext";
+
+export default function useUser() {
+    return useContext(UserContext);
+}
